+++ lib/components/TextLinkRenderer/TextLinkRenderer.docs.tsx
@@ -3,7 +3,7 @@ import { ComponentDocs } from '../../../site/src/types';
 import { TextLinkRenderer, Stack, Text, TextLink, Box } from '../';
 
 const docs: ComponentDocs = {
-  category: 'Interaction',
+  category: 'Content',
   screenshotWidths: [320],
   description: (
     <Stack space="large">