+++ lib/components/private/hideFocusRings/hideFocusRingsDataAttribute.ts
@@ -1,0 +1,1 @@
+export const hideFocusRingsDataAttribute = 'data-braid-hidefocusrings';