+++ lib/playroom/components.ts
@@ -1,6 +1,7 @@
 import '../reset';
 export * from '../components';
 
+export { AccordionItem } from '../components/AccordionItem/AccordionItem.playroom';
 export { Autosuggest } from '../components/Autosuggest/Autosuggest.playroom';
 export { ButtonLink } from '../components/ButtonLink/ButtonLink.playroom';
 export { Checkbox } from '../components/Checkbox/Checkbox.playroom';
@@ -11,6 +12,7 @@ export { Link } from '../components/Link/Link.playroom';
 export { MonthPicker } from '../components/MonthPicker/MonthPicker.playroom';
 export { Radio } from '../components/Radio/Radio.playroom';
 export { Textarea } from '../components/Textarea/Textarea.playroom';
+export { TextDropdown } from '../components/TextDropdown/TextDropdown.playroom';
 export { TextField } from '../components/TextField/TextField.playroom';
 export { TextLink } from '../components/TextLink/TextLink.playroom';
 export { Toggle } from '../components/Toggle/Toggle.playroom';