+++ lib/themes/index.ts
@@ -4,5 +4,6 @@ export { default as seekUnifiedBeta } from './seekUnifiedBeta';
 export { default as jobsDb } from './jobsDb';
 export { default as jobStreet } from './jobStreet';
 export { default as jobStreetClassic } from './jobStreetClassic';
+export { default as catho } from './catho';
 export { default as occ } from './occ';
 export { default as wireframe } from './wireframe';