+++ CHANGELOG.md
@@ -1,5 +1,177 @@
 # braid-design-system
 
+## 25.5.1
+
+### Patch Changes
+
+- Autosuggest: Don't select suggestions onBlur when using automaticSelection and suggestions are closed ([#609](https://github.com/seek-oss/braid-design-system/pull/609))
+
+## 25.5.0
+
+### Minor Changes
+
+- AccordionItem: Support `onToggle` prop without `expanded` to allow tracking in uncontrolled mode ([#605](https://github.com/seek-oss/braid-design-system/pull/605))
+
+  For example:
+
+  ```jsx
+  <AccordionItem
+    id="id"
+    label="Label"
+    onToggle={expanded => trackSomething(expanded)}
+  >
+    ...
+  </AccordionItem>
+  ```
+
+## 25.4.1
+
+### Patch Changes
+
+- Autosuggest: When using the `automaticSelection` prop, we now prevent automatic selection from ocurring if the input value hasn't changed since focusing the field ([#601](https://github.com/seek-oss/braid-design-system/pull/601))
+
+## 25.4.0
+
+### Minor Changes
+
+- Text: Add data attribute support ([#596](https://github.com/seek-oss/braid-design-system/pull/596))
+
+- Heading: Add data attribute support ([#596](https://github.com/seek-oss/braid-design-system/pull/596))
+
+## 25.3.0
+
+### Minor Changes
+
+- Inline: Add `collapseBelow` and `reverse` props. ([#593](https://github.com/seek-oss/braid-design-system/pull/593))
+
+  Similar to [Columns](https://seek-oss.github.io/braid-design-system/components/Columns), you can now responsively collapse an `Inline` into a vertical stack on mobile with the `collapseBelow` prop.
+
+  For example, if you want items to stack vertically below tablet:
+
+  ```js
+  <Inline space="small" collapseBelow="tablet">
+    ...
+  </Inline>
+  ```
+
+  Also similar to `Columns`, you can now reverse the order of items horizontally. This is particularly useful when combined with `align="right"`.
+
+  For example, if you're rendering buttons and you want your primary action on the right on desktop, but at the top on mobile:
+
+  ```js
+  <Inline space="small" collapseBelow="tablet" align="right" reverse>
+    <Button>Primary action</Button>
+    <Button weight="weak">Secondary action</Button>
+  </Inline>
+  ```
+
+- Columns: Add `align` prop ([#593](https://github.com/seek-oss/braid-design-system/pull/593))
+
+  When collapsing columns into a vertical stack on smaller screens, you can now control the alignment.
+
+  For example, if you want your columns to be horizontally centred on mobile:
+
+  ```js
+  <Columns space="small" collapseBelow="tablet" align="center">
+    <Column>...<Column>
+    <Column>...<Column>
+    <Column>...<Column>
+  </Columns>
+  ```
+
+  As a side effect, this also means that you can control the alignment of columns when the total width doesn't reach 100%.
+
+  For example:
+
+  ```js
+  <Columns space="small" align="center">
+    <Column width="1/3">...<Column>
+    <Column width="1/3">...<Column>
+  </Columns>
+  ```
+
+- Add TextDropdown component ([#594](https://github.com/seek-oss/braid-design-system/pull/594))
+
+  An inline dropdown that can be used as part of a sentence or as an
+  alternative to `Dropdown`, outside of a more structured form.
+
+  Inherits its styling from the parent typographic component, and as such
+  must be used nested within either a `Text` or `Heading` component.
+
+  Example usage:
+
+  ```tsx
+  const [jobTitle, setJobTitle] = useState('Developer');
+
+  <Text>
+    <TextDropdown
+      id="jobTitle"
+      label="Job Title"
+      value={jobTitle}
+      onChange={setJobTitle}
+      options={['Developer', 'Designer', 'Product Manager']}
+    />
+  </Text>;
+  ```
+
+### Patch Changes
+
+- Hide native focus rings on Box elements during mouse interactions ([#589](https://github.com/seek-oss/braid-design-system/pull/589))
+
+## 25.2.0
+
+### Minor Changes
+
+- Theme: Introduce the Catho theme ([#550](https://github.com/seek-oss/braid-design-system/pull/550))
+
+  Adds support to build product for the Catho market. This theme is an adaption of the [Quantum Design System](https://catho.github.io/quantum).
+
+- Add `Accordion` and `AccordionItem` components ([#587](https://github.com/seek-oss/braid-design-system/pull/587))
+
+  Example usage:
+
+  ```jsx
+  <Accordion>
+    <AccordionItem id="item_1" label="Accordion item 1">
+      <Text>Accordion item content</Text>
+    </AccordionItem>
+    <AccordionItem id="item_2" label="Accordion item 2">
+      <Text>Accordion item content</Text>
+    </AccordionItem>
+    <AccordionItem id="item_3" label="Accordion item 3">
+      <Text>Accordion item content</Text>
+    </AccordionItem>
+  </Accordion>
+  ```
+
+  Accordions manage their own state internally by default. If you'd like to take control of them yourself, you can pass an `expanded` prop to `AccordionItem` as well as an `onToggle` callback.
+
+  ```jsx
+  const [expanded, setExpanded] = useState(false);
+
+  <AccordionItem
+    id="id"
+    label="Accordion item"
+    expanded={expanded}
+    onToggle={setExpanded}
+  >
+    <Text>Accordion item content</Text>
+  </AccordionItem>;
+  ```
+
+- Box: Add support for `outline="none"` ([#587](https://github.com/seek-oss/braid-design-system/pull/587))
+
+### Patch Changes
+
+- Drop lodash usage to decrease bundle size. ([#585](https://github.com/seek-oss/braid-design-system/pull/585))
+
+  This directly affects `MonthPicker` and any components using the `data` prop:
+
+  - All field components
+  - OverflowMenu
+  - MenuRenderer
+  - Button
+
 ## 25.1.0
 
 ### Minor Changes